Feat/native tsts product 20260923

+++ src/Tsonic.CSharp.Js/TypedArrayRuntime.cs
@@ -2,14 +2,14 @@ namespace Tsonic.CSharp.Js
 {
     public static class TypedArrayRuntime
     {
-        public static double Int8BytesPerElement => 1;
-        public static double Uint8BytesPerElement => 1;
-        public static double Uint8ClampedBytesPerElement => 1;
-        public static double Int16BytesPerElement => 2;
-        public static double Uint16BytesPerElement => 2;
-        public static double Int32BytesPerElement => 4;
-        public static double Uint32BytesPerElement => 4;
-        public static double Float32BytesPerElement => 4;
-        public static double Float64BytesPerElement => 8;
+        public static int Int8BytesPerElement => 1;
+        public static int Uint8BytesPerElement => 1;
+        public static int Uint8ClampedBytesPerElement => 1;
+        public static int Int16BytesPerElement => 2;
+        public static int Uint16BytesPerElement => 2;
+        public static int Int32BytesPerElement => 4;
+        public static int Uint32BytesPerElement => 4;
+        public static int Float32BytesPerElement => 4;
+        public static int Float64BytesPerElement => 8;
     }
 }
